+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add more audience tools (create/get/delete)
+- Add contact tools
+- Add broadcast tools
+
+### Changed
+
+- Upgrade Resend SDK to v6.1.0
+- Improve `list-audiences` response formatting
+
 ## [1.1.0] - 2025-07-08
 
 ### Added